[fix/6CORE-027] fixed. Move objects from spec/.../platform/warehouse to spec/.../warehousespec.
