+++ 3-platform/warehouse/spec/src/main/java/com/_6core/platform/warehousespec/rest/v1/dto/product/ProductNameRequest.java
@@ -1,4 +1,4 @@
-package com._6core.platform.warehouse.rest.v1.dto.product;
+package com._6core.platform.warehousespec.rest.v1.dto.product;
 
 public record ProductNameRequest(Long productId, String name) {
 }
